wc: mb: Modify postcode led 3

+++ meta-facebook/wc-mb/src/platform/plat_snoop.c
@@ -11,6 +11,6 @@ void pal_postcode_led_ctl(uint8_t postcode)
 {
 	printf("******** new postcode 0x%x ********\n", postcode);
 	for (int i = 0; i < 8; i++) {
-		gpio_set(LED_POSTCODE_0 + i, (int)(postcode & BIT(i)) >> i);
+		gpio_set(LED_POSTCODE_0 + i, (postcode & BIT(i)) >> i);
 	}
 }
